move QuickList classes to a package
